tests/wpt/web-platform-tests/service-workers/service-worker/worker-interception.https.html

<!DOCTYPE html>
<title>Service Worker: intercepting Worker script loads</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>

async function setup_service_worker(t, script_url, scope) {
  const r =
      await service_worker_unregister_and_register(t, script_url, scope);
  t.add_cleanup(() => service_worker_unregister(t, scope));
  await wait_for_state(t, r.installing, 'activated');
}

promise_test(async t => {
  const worker_url = 'resources/dummy-synthesized-worker.js';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new Worker(worker_url);
  const data = await new Promise((resolve, reject) => {
    w.onmessage = e => resolve(e.data);
    w.onerror = e => reject(e.message);
  });
  assert_equals(data, 'worker loading intercepted by service worker');
}, 'Verify worker script from uncontrolled document is intercepted by ' +
   'Service Worker');

promise_test(async t => {
  const worker_url = 'resources/dummy-same-origin-worker.js';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new Worker(worker_url);
  const data = await new Promise((resolve, reject) => {
    w.onmessage = e => resolve(e.data);
    w.onerror = e => reject(e.message);
  });
  assert_equals(data, 'dummy-worker-script loaded');
}, 'Verify worker script intercepted by same-origin response succeeds');

promise_test(async t => {
  const worker_url = 'resources/dummy-cors-worker.js';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new Worker(worker_url);
  const watcher = new EventWatcher(t, w, ['message', 'error']);
  await watcher.wait_for('error');
}, 'Verify worker script intercepted by cors response fails');

promise_test(async t => {
  const worker_url = 'resources/dummy-no-cors-worker.js';
  const service_worker = 'resources/dummy-worker-interceptor.js';
  const scope = worker_url;

  await setup_service_worker(t, service_worker, scope);
  const w = new Worker(worker_url);
  const watcher = new EventWatcher(t, w, ['message', 'error']);
  await watcher.wait_for('error');
}, 'Verify worker script intercepted by no-cors cross-origin response fails');

promise_test(async t => {
  const subdoc_url = 'resources/worker-interception-iframe.https.html?bypass';
  const service_worker = 'resources/worker-load-interceptor.js';
  const scope = 'resources/';

  await setup_service_worker(t, service_worker, scope);
  const frame = await with_iframe(subdoc_url);
  t.add_cleanup(() => frame.remove());
  const data = await new Promise((resolve, reject) => {
    const channel = new MessageChannel();
    channel.port1.onmessage = e => resolve(e.data);
    frame.contentWindow.postMessage("GO", "*", [channel.port2]);
  });
  assert_equals(data.results, 'finish');
}, 'Verify worker loads from controlled document are intercepted by Service ' +
   'Worker');

</script>
</body>
